Revert "Secrets From Host Sync (#2459)"
This reverts commit 8494146a92f15bf35abb1d0995db6602cf6ed1c8.
