+++ chart/values.schema.json
@@ -3183,10 +3183,6 @@
         "configMaps": {
           "$ref": "#/$defs/EnableSwitchWithResourcesMappings",
           "description": "ConfigMaps defines if config maps in the host should get synced to the virtual cluster."
-        },
-        "secrets": {
-          "$ref": "#/$defs/EnableSwitchWithResourcesMappings",
-          "description": "Secrets defines if secrets in the host should get synced to the virtual cluster."
         }
       },
       "additionalProperties": false,
