+++ config/values.yaml
@@ -72,10 +72,6 @@ sync:
       enabled: false
     priorityClasses:
       enabled: false
-    secrets:
-      enabled: false
-      selector:
-        mappings: {}
     nodes:
       enabled: false
       syncBackChanges: false
