+++ pkg/controllers/resources/configmaps/parse_util_test.go
@@ -1,10 +1,55 @@
-package translator
+package configmaps
 
 import (
+	"sort"
 	"testing"
 
 	"k8s.io/apimachinery/pkg/types"
 )
+
+func TestParseHostNamespacesFromMappings(t *testing.T) {
+	cases := []struct {
+		name       string
+		vClusterNs string
+		mappings   map[string]string
+		expected   []string
+	}{
+		{
+			name:       "one namespace and wildcards",
+			vClusterNs: "vcluster",
+			mappings: map[string]string{
+				"":            "target/",
+				"my-ns/my-cm": "target2/my-cm",
+				"my-ns-2/*":   "target3/*",
+			},
+			expected: []string{"vcluster", "my-ns", "my-ns-2"},
+		},
+		{
+			name:       "no namespaces",
+			vClusterNs: "vcluster",
+			mappings:   map[string]string{},
+			expected:   []string{},
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			got := parseHostNamespacesFromMappings(tc.mappings, tc.vClusterNs)
+			if len(got) != len(tc.expected) {
+				t.Logf("expectedVirtual %d namespaces, got %d", len(tc.expected), len(got))
+				t.Fail()
+			}
+			sort.Strings(got)
+			sort.Strings(tc.expected)
+			for i := range got {
+				if tc.expected[i] != got[i] {
+					t.Logf("expectedVirtual %s, got %s", tc.expected[i], got[i])
+					t.Fail()
+				}
+			}
+		})
+	}
+}
 
 func TestMatches(t *testing.T) {
 	cases := []struct {
@@ -98,9 +143,7 @@ func TestMatches(t *testing.T) {
 	t.Run("match host", func(t *testing.T) {
 		for _, tc := range cases {
 			t.Run(tc.name, func(t *testing.T) {
-				got, _ := matchesHostObject(tc.hostName, tc.hostNs, tc.mappings, "vcluster", func(hostName, _ string) bool {
-					return hostName == "kube-root-ca.crt"
-				})
+				got, _ := matchesHostObject(tc.hostName, tc.hostNs, tc.mappings, "vcluster", skipKubeRootCaConfigMap)
 				if got.Name == tc.virtualName && got.Namespace == tc.virtualNs {
 					return
 				}
