+++ pkg/syncer/types/translator.go
@@ -2,7 +2,6 @@ package types
 
 import (
 	syncercontext "github.com/loft-sh/vcluster/pkg/syncer/synccontext"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
 )
 
@@ -18,12 +17,4 @@ type GenericTranslator interface {
 
 	// EventRecorder returns
 	EventRecorder() record.EventRecorder
-}
-
-type FromConfigTranslator interface {
-	Translator
-	// EventRecorder returns
-	EventRecorder() record.EventRecorder
-	MatchesHostObject(hostName, hostNamespace string) (types.NamespacedName, bool)
-	MatchesVirtualObject(virtualName, virtualNamespace string) (types.NamespacedName, bool)
 }
