+++ test/e2e/values.yaml
@@ -40,10 +40,4 @@ sync:
         mappings:
           "from-host-sync-test/*": "barfoo/*"
           "default/my-cm": "barfoo/cm-my"
-          "": "default"
-    secrets:
-      enabled: true
-      selector:
-        mappings:
-          "from-host-sync-test-2/dummy": "barfoo2/dummy"
-          "default/my-secret": "barfoo2/secret-my"
+          "": "my-new-ns"
